requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
pymongo==4.7.3
python-dotenv==1.0.1
umap-learn==0.5.6
hdbscan==0.8.38
requests==2.32.3
pandas==2.2.2
transformers==4.42.4
sentence-transformers==3.0.1
gunicorn==22.0.0
eventlet==0.36.1
numpy==1.26.4
scikit-learn==1.5.0